eval_framework/requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0,<3.0
python-dotenv>=1.0
httpx>=0.25.0

# Database drivers (install only what you need, but we list both)
asyncpg>=0.27.0        # For PostgreSQL
motor>=3.0             # For MongoDB